move events and payload logic into it's own folder
