+++ rust/capture/src/events/mod.rs
@@ -1,0 +1,6 @@
+pub mod analytics;
+pub mod recordings;
+
+// Re-export commonly used types
+pub use analytics::{process_events, process_single_event};
+pub use recordings::{process_replay_events, RawRecording, RecordingRequest};
